Fix spurious error highlighting on GitHub.com (#1)
When rendered using Lightshow (https://github-lightshow.herokuapp.com/),
a valid line such as `00:00.000 --> 01:45.000` will be highlighted as an
error. This is because the pattern that was *supposed* to match used `^`
to match the beginning of the containing pattern block. Atom incorrectly
treats ^ and \G as synonymous with respect to `#include`'d patterns, but
GitHub (correctly) matches ^ only at the beginning of a line/string.

As a result, the #timestamp pattern failed to match a trailing timestamp
inside `-->\s+(.+?)$`, leaving it to fall to the "match anything else as
invalid" rule that followed.